refactor: extract AppModel into app_model module
- update.rs mixed two concerns: pure message dispatch logic and the IO
  state machine (AppModel with its file-saving methods)
- AppModel now lives in app_model.rs, re-exported from update for
  backward compat (all update::AppModel imports continue to work)
- update.rs is left with UpdateResult, UpdateError, update(), and its
  pure helper functions only
