+++ crates/am/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod alias;
 pub mod alias_id;
+pub mod app_model;
 pub mod cli;
 pub mod config;
 pub mod dirs;
